Change `bestpokeApplyOffline` test to `wasteActivityApplyOffline` test
The original test was generic and applied to all routes with the `bespokeApplyOffline` controller. Here we rework it to apply only to waste activities and be more explicit about what it tests
