--- test/routes/bespokeApplyOffline.route.test.js
@@ -1,149 +1,0 @@
-'use strict'
-
-const Lab = require('@hapi/lab')
-const lab = exports.lab = Lab.script()
-const Code = require('@hapi/code')
-const sinon = require('sinon')
-const Mocks = require('../helpers/mocks')
-const GeneralTestHelper = require('./generalTestHelper.test')
-
-const server = require('../../server')
-const Application = require('../../src/persistence/entities/application.entity')
-const RecoveryService = require('../../src/services/recovery.service')
-const LoggingService = require('../../src/services/logging.service')
-const CookieService = require('../../src/services/cookie.service')
-const { COOKIE_RESULT } = require('../../src/constants')
-const ItemEntity = require('../../src/persistence/entities/item.entity')
-const Routes = require('../../src/routes')
-
-const routesToTest = {}
-
-// Iterate over Routes and add any routes with the controller bespokeApplyOffline to the routesToTest object. We have a
-// deny list of routes which have this controller but which do not have behaviour defined in
-// BespokeApplyOfflineController. These routes should be removed from the deny list if and when the controller is
-// updated to act on them.
-const routeDenyList = ['FACILITY_APPLY_OFFLINE', 'WASTE_ASSESSMENT_APPLY_OFFLINE']
-
-for (const key in Routes) {
-  const item = Routes[key]
-  if (item.controller === 'bespokeApplyOffline' && !routeDenyList.includes(key)) {
-    routesToTest[key] = item
-  }
-}
-
-const ACTIVITY_ITEMS = [
-  {
-    id: 'ITEM_1',
-    itemName: 'ITEM_1',
-    shortName: 'item-1',
-    canApplyFor: true,
-    canApplyOnline: true
-  }, {
-    id: 'ITEM_2',
-    itemName: 'ITEM_2',
-    shortName: 'item-2',
-    canApplyFor: true,
-    canApplyOnline: false
-  }, {
-    id: 'ITEM_3',
-    itemName: 'ITEM_3',
-    shortName: 'item-3',
-    canApplyFor: false,
-    canApplyOnline: true
-  }, {
-    id: 'ITEM_4',
-    itemName: 'ITEM_4',
-    shortName: 'item-4',
-    canApplyFor: false,
-    canApplyOnline: false
-  }
-]
-
-const comparisonValues = ACTIVITY_ITEMS
-  .filter((item) => !item.canApplyFor || !item.canApplyOnline)
-  .map(({ id }) => id)
-
-Object.entries(routesToTest).forEach(([route, { path: routePath, itemType, pageHeading, pageDescription, nextRoute }]) => {
-  lab.experiment(route, () => {
-    let sandbox
-    let mocks
-
-    lab.beforeEach(() => {
-      mocks = new Mocks()
-
-      mocks.taskDeterminants.wasteActivities = ['item-1', 'item-2', 'item-3', 'item-4']
-
-      // Create a sinon sandbox to stub methods
-      sandbox = sinon.createSandbox()
-
-      // Stub methods
-      sandbox.stub(CookieService, 'validateCookie').value(() => COOKIE_RESULT.VALID_COOKIE)
-      sandbox.stub(Application.prototype, 'isSubmitted').value(() => false)
-      sandbox.stub(ItemEntity, 'listWasteActivitiesForFacilityTypes').value(() => ACTIVITY_ITEMS)
-      sandbox.stub(RecoveryService, 'createApplicationContext').value(() => mocks.recovery)
-    })
-
-    lab.afterEach(() => {
-      // Restore the sandbox to make sure the stubs are removed correctly
-      sandbox.restore()
-    })
-
-    lab.experiment('bespokeApplyOffline page tests:', () => {
-      new GeneralTestHelper({ lab, routePath }).test({
-        excludeCookiePostTests: true
-      })
-
-      lab.experiment(`GET ${routePath}`, () => {
-        let getRequest
-
-        lab.beforeEach(() => {
-          getRequest = {
-            method: 'GET',
-            url: routePath,
-            headers: {}
-          }
-        })
-
-        lab.test('success', async () => {
-          const doc = await GeneralTestHelper.getDoc(getRequest)
-          Code.expect(doc.getElementById('page-heading').firstChild.nodeValue).to.equal(pageHeading)
-
-          // Test for the existence of expected static content
-          GeneralTestHelper.checkElementsExist(doc, [
-            'change-selection-link',
-            'how-to-apply',
-            'bespoke-prefix',
-            'bespoke-link',
-            'bespoke-text'
-          ])
-        })
-
-        lab.experiment('failure', () => {
-          lab.test('redirects to error screen when failing to recover the application', async () => {
-            const spy = sandbox.spy(LoggingService, 'logError')
-            RecoveryService.createApplicationContext = () => {
-              throw new Error('recovery failed')
-            }
-
-            const res = await server.inject(getRequest)
-            Code.expect(spy.callCount).to.equal(1)
-            Code.expect(res.statusCode).to.equal(500)
-          })
-        })
-
-        lab.experiment('checks', () => {
-          lab.test('correctly displays names of permits that can\'t be applied for', async () => {
-            const doc = await GeneralTestHelper.getDoc(getRequest)
-
-            const values = Object
-              .values(doc.getElementById('items').childNodes)
-              .filter(({ firstChild }) => firstChild && firstChild.nodeValue)
-              .map(({ firstChild }) => firstChild.nodeValue)
-
-            Code.expect(values).to.only.include(comparisonValues)
-          })
-        })
-      })
-    })
-  })
-})
